dom/plugins/test/mochitest/test_fullpage.html

<head>
  <title>Full-page seekable stream</title>

  <script type="text/javascript"
          src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="plugin-utils.js"></script>
  <link rel="stylesheet" type="text/css"
        href="/tests/SimpleTest/test.css">

<body>
  <p id="display"></p>

  <script type="application/javascript">
    SimpleTest.waitForExplicitFinish();
    setTestPluginEnabledState(SpecialPowers.Ci.nsIPluginTag.STATE_ENABLED);

    function frameLoaded() {
      var testframe = document.getElementById('testframe');
      var content = testframe.contentDocument.body.innerHTML;
      if (!content.length)
        return;

      var req = new XMLHttpRequest();
      req.open('GET', 'loremipsum.xtest', false);
      req.overrideMimeType('text/plain; charset=x-user-defined');
      req.send(null);
      is(req.status, 200, "bad XMLHttpRequest");
      is(content, req.responseText.replace(/\r\n/g, "\n"),
         "content doesn't match");
      SimpleTest.finish();
    }
  </script>

  <iframe src="loremipsum.xtest" streamtype="seek"></iframe>
  <iframe id="testframe" name="testframe" onload="frameLoaded()"></iframe>
